+++ src/rpc/rpc_util.rs
@@ -41,7 +41,12 @@ pub fn is_streaming_method(method_name: &str) -> bool {
     STREAMING_METHODS.contains(&method_name)
 }
 
-const V1_METHODS: [&str; 2] = [ETH_CHAIN_ID, ETH_GET_BALANCE];
+const V1_METHODS: [&str; 4] = [
+    ETH_ACCOUNTS,
+    ETH_BLOCK_NUMBER,
+    ETH_CHAIN_ID,
+    ETH_GET_BALANCE,
+];
 
 pub fn is_v1_method(method_name: &str) -> bool {
     V1_METHODS.contains(&method_name)