+++ src/rpc_client/eth_ops.rs
@@ -8,6 +8,14 @@ use crate::eth::{Address, BlockNumberOrHash};
 use super::{ApiInfo, RpcRequest};
 
 impl ApiInfo {
+    pub fn eth_accounts_req() -> RpcRequest<Vec<String>> {
+        RpcRequest::new_v1(ETH_ACCOUNTS, ())
+    }
+
+    pub fn eth_block_number_req() -> RpcRequest<String> {
+        RpcRequest::new_v1(ETH_BLOCK_NUMBER, ())
+    }
+
     pub fn eth_chain_id_req() -> RpcRequest<String> {
         RpcRequest::new_v1(ETH_CHAIN_ID, ())
     }