+++ src/state_manager/mod.rs
@@ -41,7 +41,7 @@ use chain_rand::ChainRand;
 use cid::Cid;
 
 use fil_actor_interface::miner::SectorOnChainInfo;
-use fil_actor_interface::miner::{MinerInfo, MinerPower};
+use fil_actor_interface::miner::{MinerInfo, MinerPower, Partition};
 use fil_actor_interface::*;
 use fil_actors_shared::fvm_ipld_amt::Amtv0 as Amt;
 use fil_actors_shared::fvm_ipld_bitfield::BitField;
@@ -1047,26 +1047,43 @@ where
         addr: &Address,
         ts: &Arc<Tipset>,
     ) -> Result<BitField, Error> {
+        self.all_partition_sectors(addr, ts, |partition| partition.faulty_sectors().clone())
+    }
+    /// Retrieves miner recoveries.
+    pub fn miner_recoveries(
+        self: &Arc<Self>,
+        addr: &Address,
+        ts: &Arc<Tipset>,
+    ) -> Result<BitField, Error> {
+        self.all_partition_sectors(addr, ts, |partition| partition.recovering_sectors().clone())
+    }
+
+    fn all_partition_sectors(
+        self: &Arc<Self>,
+        addr: &Address,
+        ts: &Arc<Tipset>,
+        get_sector: impl Fn(Partition<'_>) -> BitField,
+    ) -> Result<BitField, Error> {
         let actor = self
             .get_actor(addr, *ts.parent_state())?
             .ok_or_else(|| Error::State("Miner actor not found".to_string()))?;
 
         let state = miner::State::load(self.blockstore(), actor.code, actor.state)?;
 
-        let mut faults = Vec::new();
+        let mut partitions = Vec::new();
 
         state.for_each_deadline(
             &self.chain_config.policy,
             self.blockstore(),
             |_, deadline| {
                 deadline.for_each(self.blockstore(), |_, partition| {
-                    faults.push(partition.faulty_sectors().clone());
+                    partitions.push(get_sector(partition));
                     Ok(())
                 })
             },
         )?;
 
-        Ok(BitField::union(faults.iter()))
+        Ok(BitField::union(partitions.iter()))
     }
 
     /// Retrieves miner power.