+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 
 ### Added
 
+- [#3779](https://github.com/ChainSafe/forest/pull/3779) Implement the
+  `Filecoin.StateMinerRecoveries` lotus-compatible RPC API.
 - [#3745](https://github.com/ChainSafe/forest/pull/3745) Implement the
   `Filecoin.StateCirculatingSupply` lotus-compatible RPC API.
 - [#3773](https://github.com/ChainSafe/forest/pull/3773) Implement the
@@ -38,6 +40,8 @@
   `forest-tool snapshot compute-state` and `Filecoin.StateCall` RPC API.
 - [#3720](https://github.com/ChainSafe/forest/pull/3750) Implement the
   `Filecoin.StateMinerInfo` lotus-compatible RPC API.
+- [#1670](https://github.com/ChainSafe/forest/issues/1670) Support Butterflynet
+  🦋.
 
 ### Changed